requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.146.0
pandas==2.2.2
openpyxl==3.1.5
firebase-admin==6.5.0
email-validator
